+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This list is not currently intended to be all-encompassing - it will document major and breaking API changes with their rationale when appropriate:
 
+### v3.125.0 (uncut)
+- Added `http4k-testing-approval` module, which is compatible with JUnit5 tests and is compatible with 
+the `OkeyDoke` approval testing files and IntelliJ plugin.
+
 ### v3.124.0
 - Make authentication response available when creating AuthorizationCode